requirements.txt
Django==2.1.2
django-debug-toolbar==1.10.1
djangorestframework==3.8.2
pytz==2018.5
sqlparse==0.2.4
